+++ pkg/go/graph/weighted_graph.go
@@ -34,63 +34,6 @@ func NewWeightedAuthorizationModelGraph(model *openfgav1.AuthorizationModel) (*W
 	return wg, nil
 }
 
-// nolint: cyclop
-func NewWeightedAuthorizationModelGraphBuilder(model *openfgav1.AuthorizationModel) (*WeightedAuthorizationModelGraphBuilder, error) {
-	g, err := NewAuthorizationModelGraph(model)
-	if err != nil {
-		return nil, err
-	}
-	g, err = g.Reversed() // we want edges to have the direction of Check when doing the weight assignments later
-	if err != nil {
-		return nil, err
-	}
-
-	graphBuilder := &WeightedAuthorizationModelGraphBuilder{multi.NewDirectedGraph(), DrawingDirectionCheck}
-
-	// Add all nodes
-	iterNodes := g.Nodes()
-	for iterNodes.Next() {
-		nextNode := iterNodes.Node()
-		node, ok := nextNode.(*AuthorizationModelNode)
-		if !ok {
-			return nil, fmt.Errorf("%w: could not cast to WeightedAuthorizationModelNode", ErrBuildingGraph)
-		}
-		newNode := NewWeightedAuthorizationModelNode(node, false)
-		graphBuilder.AddNode(newNode)
-	}
-
-	// Add all the edges
-	iterEdges := g.Edges()
-	for iterEdges.Next() {
-		nextEdge, ok := iterEdges.Edge().(multi.Edge)
-		if !ok {
-			return nil, fmt.Errorf("%w: could not cast %v to multi.Edge", ErrBuildingGraph, iterEdges.Edge())
-		}
-
-		// NOTE: because we use a multigraph, one edge can include multiple lines, so we need to add each line individually.
-		iterLines := nextEdge.Lines
-		for iterLines.Next() {
-			nextLine := iterLines.Line()
-			castedEdge, ok := nextLine.(*AuthorizationModelEdge)
-			if !ok {
-				return nil, fmt.Errorf("%w: could not cast %v to AuthorizationModelEdge", ErrBuildingGraph, nextLine)
-			}
-
-			err = graphBuilder.AddEdgeAndUpdateNodes(castedEdge)
-			if err != nil {
-				return nil, err
-			}
-		}
-	}
-
-	err = graphBuilder.AssignWeights()
-	if err != nil {
-		return nil, err
-	}
-
-	return graphBuilder, nil
-}
-
 var _ dot.Attributers = (*WeightedAuthorizationModelGraph)(nil)
 
 func (wg *WeightedAuthorizationModelGraph) DOTAttributers() (encoding.Attributer, encoding.Attributer, encoding.Attributer) {
